part 4 ready

+++ controllers/blogs.js
@@ -1,8 +1,14 @@
 const blogsRouter = require('express').Router()
 const BlogModel = require('../models/blog')
+const UserModel = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await BlogModel.find({})
+  // console.log(request)
+
+  const blogs = await BlogModel
+    .find({}).populate('user', { username: 1, name: 1 })
 
   response.json(blogs.map(blog => blog.toJSON()))
 })
@@ -35,19 +41,29 @@ blogsRouter.get('/:id', async (request, response, next) => {
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
 
-  if (body.title === undefined || body.url === undefined) {
-    return response.status(400).json({ error: 'title or url missing' })
-  }
-
-  const blogModel = new BlogModel({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0
-  })
-
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await UserModel.findById(decodedToken.id)
+
+    if (body.title === undefined || body.url === undefined) {
+      return response.status(400).json({ error: 'title or url missing' })
+    }
+
+    const blogModel = new BlogModel({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user._id
+    })
+
     const savedBlog = await blogModel.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     const savedAndFormattedBlog = savedBlog.toJSON()
     response.status(201).json(savedAndFormattedBlog)
   } catch (exception) {
@@ -66,7 +82,23 @@ blogsRouter.post('/', async (request, response, next) => {
 */
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    // console.log(request.param.id)
+
+    const blog = await BlogModel.findById(request.params.id)
+    // console.log(blog)
+    // console.log(decodedToken.id)
+
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: 'wrong user' })
+    }
+
     await BlogModel.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
